backend/requirements.txt
django
djangorestframework
drf-nested-routers
channels[daphne]